refactor: change CustomDatePickerViwe class to UpdateItemView  extension
